Add vhost user SCMI support
SCMI is a set of operating system-independent software interfaces
that are used in system management.

vhost user SCMI include two parts, one is the vhost user frontend, it is
in crosvm-gunyah + crosvm, the other is the vhost user backend, it is in
a dedicated process.

This patch only includes the part of vhost user frontend in
crosvm-gunyah.

Change-Id: I9acf6d98d34b73724ffec171d16578bdee1fe1dd
Signed-off-by: Conghui Chen <quic_conghuic@quicinc.com>