qcrosvm: Add vhost user fastrpc support in qcrosvm
vhost-user fastrpc provides a virtual fastrpc protocol device to
the Guest VM. This patch only includes the part of vhost
user frontend in qcrosvm, providing a fastrpc protocol mmio device
based on gunyah-driver to gvm.

It connects to vhost-User fastrpc backend over a socket to do negotiation of features, config-space, Irqfd , Eventfd & Virtqueue memory descriptors.
The socket and label info is provided to qcrosvm in the follwoing format:
--vhost-user-frpc <socket_path>,label=<label_no>

Change-Id: I674d8edc7b2b07e871d1a090a24efc62911f00f2