qcrosvm: Add initialization support for the vHost User Glink over CMA device
The vHost User Glink Passthrough protocol is used for communicating CMA information between the driver and device.
This CMA information is used by the driver, in the Guest VM, to establish the Glink protocol with remote subsystems.

This change includes the initialization of the VMM portion of the vHost User Glink Passthrough protocol as an MMIO device.
The device portion of this protocol resides in a separate process.

Change-Id: I63f48f08340766ba0eb1e5d307e58b6030320dab