qcrosvm: main: Add support for non-protected VirtIO communication
Add support to query a VMs' number of memory regions, and the ability to
query each region individually for information needed to build the
Guest memory FD.

The Guest memory FDs are used by devices running in the VMM, vHost, or
vHost User context to access the virtqueues and data buffers created by
the driver.

Introduce a command line argument "--use-non-protected-virtio" to
optionally enable traditional VirtIO, as opposed to a
non-traditional/bounce buffer enabled VirtIO. When the option is set,
Guest memory FDs are created out of the entire Guest memory region.
When the option is not set, only an explicit shared region of memory
is used to create the Guest memory FD.

Change-Id: I73e28553fe51dc570ace5883a183b96f64ade2f0
Signed-off-by: Anant Goel <quic_anantg@quicinc.com>